// apps/webapp/src/shared/api/generated/connectors/api.ts
export interface ConnectionCreateDto {
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'username': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'orgId'?: string;
    /**
     * 
     * @type {object}
     * @memberof ConnectionCreateDto
     */
    'type'?: object;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'agentId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'runtimeEnvironmentId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'instanceDisplayName': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'host': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'port': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'database': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'codepage': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'createdBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'updatedBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'adjustedJdbcHostName': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'shortDescription': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'password': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'createTime': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionCreateDto
     */
    'updateTime': string;
}
/**
 * 
 * @export
 * @interface ConnectionEntity
 */
export interface ConnectionEntity {
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'orgId': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'type': ConnectionEntityTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'description': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'agentId': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'runtimeEnvironmentId': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'instanceDisplayName': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'host': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'database': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'codepage': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'adjustedJdbcHostName': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'shortDescription': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'port': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'username': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'password': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'createTime': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'updateTime': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'createdBy': string;
    /**
     * 
     * @type {string}
     * @memberof ConnectionEntity
     */
    'updatedBy': string;
}

export const ConnectionEntityTypeEnum = {
    CSV_FILE: 'CSVFile',
    MY_SQL: 'MySQL',
    AMAZON_S3_V2: 'Amazon_S3_v2',
    ORACLE: 'Oracle',
    FTP: 'FTP',
    SALESFORCE: 'Salesforce',
    MS_ACCESS: 'MS_ACCESS',
    WEB_SERVICES_CONSUMER: 'WebServicesConsumer',
    MSD: 'MSD',
    SQL_SERVER: 'SqlServer',
    SNOWFLAKE_CLOUD_DATA_WAREHOUSE_V2: 'Snowflake_Cloud_Data_Warehouse_V2',
    TOOLKIT_CCI: 'TOOLKIT_CCI'
} as const;

export type ConnectionEntityTypeEnum = typeof ConnectionEntityTypeEnum[keyof typeof ConnectionEntityTypeEnum];